# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "описание сервиса",
        "title": "testProject",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Автор",
            "email": "Имейл"
        },
        "version": "0.1"
    },
    "host": "localhost:9091",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/user": {
            "post": {
                "description": "Создание пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Создание пользователя",
                "operationId": "CreateOrder",
                "parameters": [
                    {
                        "description": "Входящие данные",
                        "name": "rq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CreateResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateResp": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.UserCreateReq": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pass": {
                    "type": "string"
                },
                "surName": {
                    "type": "string"
                }
            }
        }
    }
}